feat: improve datastructure types
